// api_v2/src/enums/notification-type.enum.ts
export enum NotificationTypeEnum {
  LEAVE = 'leave',
  OVERTIME = 'overtime',
  OVERTIME_SUMMARY = 'overtime summary',
  UNDERTIME = 'undertime',
  CHANGE_SHIFT = 'change shift',
  ESL_OFFSET_SCHEDULE = 'esl change shift',
  ESL_OFFSET = 'offset',
  CHANGE_SCHEDULE = 'change schedule',
  LEAVE_RESOLVED = 'leave_resolved',
  OVERTIME_RESOLVED = 'overtime_resolved',
  UNDERTIME_RESOLVED = 'undertime_resolved',
  CHANGE_SHIFT_RESOLVED = 'change shift_resolved',
  ESL_OFFSET_SCHEDULE_RESOLVED = 'esl change shift_resolved',
  ESL_OFFSET_RESOLVED = 'offset_resolved',
  CHANGE_SCHEDULE_RESOLVED = 'change schedule_resolved',
}